docs: deprecate Ventoy recommendation — it’s insecure
Ventoy uses binary blobs which can't be trusted to be free of malware or compliant to their licenses.
It’s also flagged as unfree and insecure in nixpkgs, so it’s no longer safe for creating bootable USBs.

See these issues for details:

- https://github.com/ventoy/Ventoy/issues/2795
- https://github.com/ventoy/Ventoy/issues/3224

Stick with the simple, transparent, and more secure `dd` command instead.